+++ src/components/vetdoc/ReservacionesPendientes.tsx
@@ -72,6 +72,7 @@ export const ReservacionesPendientes: React.FC<ReservacionesPendientesProps> = (
                                 </div>
                                 <div className="text-sm">
                                     <span className="font-medium">Estado:</span> {reservacion.Estado}
+                                    <span className="font-medium"> - Motivo:</span> {reservacion.Motivo}
                                 </div>
                             </div>
                             <Button
